[misc] support r2dbc-spi 0.9.1 spec
0.9.1-spec variant: r2dbc-spi 0.9.1.RELEASE, artifactId r2dbc-mariadb-0.9.1-spec, getRowsUpdated() returns Integer, RowMetadata.getColumnNames() override, adapted test assertions.

+++ src/main/java/org/mariadb/r2dbc/client/MariadbSegmentResult.java
@@ -160,7 +160,7 @@ public final class MariadbSegmentResult extends AbstractReferenceCounted impleme
   }
 
   @Override
-  public Mono<Long> getRowsUpdated() {
+  public Mono<Integer> getRowsUpdated() {
     return this.segments
         .<Integer>handle(
             (segment, sink) -> {
@@ -185,7 +185,7 @@ public final class MariadbSegmentResult extends AbstractReferenceCounted impleme
                 return;
               }
 
-              long sum = 0;
+              int sum = 0;
 
               for (Integer integer : list) {
                 sum += integer;
